Added time checking for password checking optimization. Fix in description.
Previously, incorrect password checks resulted in locking out the account.
This was fixed; however, I wanted to keep an account locked out if it was before checking it. This has been implemented. Regardless of its locked out state, a checked account with be locked out/not locked out after checking depending on how it was before.